Try to restore older cache if the previous is missing.
This approach should fix edge case when developer create a new build during the day. Then it will create a new build which he cancel. And after that will try to create a 3rd build. In the old version it will create a new changelog with features which were previously builded that day. With this approach it should not since it should hit the cache created that day.